+++ NEWS.md
@@ -1,3 +1,7 @@
+# bstfun (development version)
+
+* Added new function `bold_italicize_group_labels()` to bold and italicize group header labels in gtsummary tables. (#65)
+
 # bstfun 0.3.0
 
 * Function `add_cuminc_risktables()` now allows for 3 groups. Makes colors consistent between labels and estimates.  Removed "Number at risk" label and places a label ", n:" next to each strata name.