+++ NEWS.md
@@ -1,5 +1,6 @@
-# bstfun (development version)
+# bstfun 0.3.1
 
+* Updates to the templating files.
 
 * Updated custom template definitions to utilize parts from the default template, reducing the amount of code needed to create a custom template.